Split the velocity-advection stencils into three modules
compute_velocity_advection.py had grown to 1288 lines holding both
programs and all 19 operators. It now splits along the only line that
matters here, which of the two programs reaches a definition:

  velocity_advection_terms.py      15 step-agnostic operators
  velocity_advection_predictor.py  the predictor program + its 3 operators
  velocity_advection_corrector.py  the corrector program + its 1 operator

The stencil tests are arranged the same way, with the numpy references
in test_velocity_advection_terms.py where both step files import them.

The partition is only this clean because the vertical-momentum operator
was unified first; before that the two steps had diverged versions and
the terms module would not have been shared.
